Correction Sandwich

+++ src/main/java/boundary/SandwichResource.java
@@ -4,12 +4,15 @@ import entity.Bread;
 import entity.Ingredient;
 import entity.Sandwich;
 import entity.SandwichBindIngredientsAndBread;
+import exception.SandwichBadRequest;
+import exception.SandwichNotFoundExeception;
 
 import javax.ejb.Stateless;
 import javax.jms.Session;
 import javax.persistence.*;
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.NotFoundException;
+import javax.ws.rs.core.NoContentException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -23,15 +26,15 @@ public class SandwichResource {
     @PersistenceContext
     EntityManager em;
 
-    public Sandwich create(SandwichBindIngredientsAndBread s) throws Exception {
+    public Sandwich create(SandwichBindIngredientsAndBread s) throws EntityNotFoundException, BadRequestException, SandwichBadRequest {
         Sandwich res = new Sandwich();
         res.setId(UUID.randomUUID().toString());
         int tailleSandwich = s.getTaille();
 
-        if (tailleSandwich > 4 || tailleSandwich < 7) {
+        if (tailleSandwich >= 4 && tailleSandwich <= 7) {
             res.setTaille(tailleSandwich);
         } else {
-            throw (new BadRequestException());
+            throw new SandwichBadRequest("La Taille doit etre comprise entre 7 et 9");
         }
 
 
@@ -46,7 +49,7 @@ public class SandwichResource {
         if (b != null) {
             res.setBread(b);
         } else {
-            throw (new BadRequestException());
+            throw new SandwichBadRequest("ID pain inexistant");
         }
         //get all id in list of id ingredients and add ingredients to sandwich
         List<Ingredient> listIng = new ArrayList<Ingredient>();
@@ -55,7 +58,7 @@ public class SandwichResource {
             if (s.getIdIngredients().get(i) != null) {
                 listIng.add(this.em.find(Ingredient.class, s.getIdIngredients().get(i)));
             } else {
-                throw (new BadRequestException());
+                throw new SandwichBadRequest("Un id d'ingredients n'existe pas");
             }
         }
 
@@ -106,55 +109,73 @@ public class SandwichResource {
         return this.em.merge(res);
     }
 
-    public List<Sandwich> findAll() {
+    public List<Sandwich> findAll() throws NoContentException{
         Query q = this.em.createNamedQuery("Sandwich.findAll", Sandwich.class);
         // pour éviter les pbs de cache
         q.setHint("javax.persistence.cache.storeMode", CacheStoreMode.REFRESH);
-        return q.getResultList();
+        List<Sandwich> res = q.getResultList();
+        if(res.size() != 0){
+            return res;
+        }else{
+            throw (new NoContentException("Pas de sandwich pour le moment"));
+        }
+
     }
 
 
-    public Sandwich findById(String id) throws EntityNotFoundException {
+    public Sandwich findById(String id) throws SandwichNotFoundExeception {
         Sandwich s = this.em.find(Sandwich.class, id);
+        if(s == null){
+            throw  new SandwichNotFoundExeception("Pas de sandwich avec cette id");
+        }
         return s;
     }
 
 
-    public Sandwich update(SandwichBindIngredientsAndBread s, String id) throws EntityNotFoundException, BadRequestException {
-        Sandwich ref = this.em.find(Sandwich.class, id);
+    public Sandwich update(SandwichBindIngredientsAndBread s, String id) throws SandwichBadRequest , SandwichNotFoundExeception {
+            Sandwich ref = this.em.find(Sandwich.class, id);
+            if (ref == null) throw (new SandwichNotFoundExeception("Pas de sandwich avec cette id"));
 
-        int tailleSandwich = s.getTaille();
-        if (tailleSandwich > 4 || tailleSandwich < 7) {
-            ref.setTaille(tailleSandwich);
-        } else {
-            throw (new BadRequestException());
-        }
 
-        //get bread by id and add to sandwich
-        Bread b = this.em.find(Bread.class, s.getIdBread());
-        if (b != null) {
-            ref.setBread(b);
-        }
-        //get all id in list of id ingredients and add ingredients to sandwich
-        List<Ingredient> listIng = new ArrayList<Ingredient>();
-
-        for (int i = 0; i < s.getIdIngredients().size(); i++) {
-           String val =  s.getIdIngredients().get(i);
-            if (val != null) {
-                listIng.add(this.em.find(Ingredient.class, val));
-            }else{
-                throw  (new BadRequestException());
+            int tailleSandwich = s.getTaille();
+            if (tailleSandwich >= 4 && tailleSandwich <= 7) {
+                ref.setTaille(tailleSandwich);
+            } else {
+                throw new SandwichBadRequest("La taille doit etre comprise entre  4 et 7");
             }
+
+
+            //get bread by id and add to sandwich
+            Bread b = this.em.find(Bread.class, s.getIdBread());
+            if (b != null) {
+                ref.setBread(b);
+            } else {
+                throw new SandwichBadRequest("ID pain inexistant");
+            }
+            //get all id in list of id ingredients and add ingredients to sandwich
+            List<Ingredient> listIng = new ArrayList<Ingredient>();
+
+            for (int i = 0; i < s.getIdIngredients().size(); i++) {
+                String val = s.getIdIngredients().get(i);
+                if (val != null) {
+                    listIng.add(this.em.find(Ingredient.class, val));
+                } else {
+                    throw new SandwichBadRequest("Un id d'ingredients n'existe pas");
+                }
+            }
+
+            //Category ref = this.em.getReference(Category.class, id);
+            ref.setIngredients(listIng);
+
+            return this.em.merge(ref);
         }
 
-        //Category ref = this.em.getReference(Category.class, id);
-        ref.setIngredients(listIng);
 
-        return this.em.merge(ref);
-    }
-
-    public void delete(String id) throws EntityNotFoundException {
-        Sandwich ref = this.em.getReference(Sandwich.class, id);
+    public void delete(String id) throws SandwichNotFoundExeception {
+        Sandwich ref = this.em.find(Sandwich.class, id);
+        if(ref == null){
+            throw  new SandwichNotFoundExeception("id de sandwich non existant");
+        }
         this.em.remove(ref);
     }
 }
